+++ eras/allegra/impl/src/Cardano/Ledger/Allegra/UTxO.hs
@@ -15,7 +15,6 @@ import Cardano.Ledger.Shelley.UTxO (
   getShelleyMinFeeTxUtxo,
   getShelleyScriptsNeeded,
   getShelleyWitsVKeyNeeded,
-  shelleyConsumed,
   shelleyProducedValue,
  )
 import Cardano.Ledger.State (EraUTxO (..), ScriptsProvided (..))
@@ -24,9 +23,7 @@ import Lens.Micro
 instance EraUTxO AllegraEra where
   type ScriptsNeeded AllegraEra = ShelleyScriptsNeeded AllegraEra
 
-  consumed = shelleyConsumed
-
-  getConsumedValue pp lookupKeyDeposit _ = getConsumedCoin pp lookupKeyDeposit
+  getConsumedValue pp lookupKeyDeposit = getConsumedCoin pp lookupKeyDeposit
 
   getProducedValue = shelleyProducedValue