+++ eras/conway/impl/src/Cardano/Ledger/Conway/UTxO.hs
@@ -10,7 +10,6 @@
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module Cardano.Ledger.Conway.UTxO (
-  conwayConsumed,
   conwayProducedValue,
   getConwayWitsVKeyNeeded,
   getConwayScriptsNeeded,
@@ -105,19 +104,6 @@ getConwayScriptsNeeded utxo txBody =
             TreasuryWithdrawals _ (SJust guardrailsScriptHash) -> Just guardrailsScriptHash
             _ -> Nothing
 
-conwayConsumed ::
-  (EraUTxO era, ConwayEraCertState era) =>
-  PParams era ->
-  CertState era ->
-  UTxO era ->
-  TxBody l era ->
-  Value era
-conwayConsumed pp certState =
-  getConsumedValue
-    pp
-    (lookupDepositDState $ certState ^. certDStateL)
-    (lookupDepositVState $ certState ^. certVStateL)
-
 conwayProducedValue ::
   ( ConwayEraTxBody era
   , Value era ~ MaryValue
@@ -132,8 +118,6 @@ conwayProducedValue pp isStakePool txBody =
 
 instance EraUTxO ConwayEra where
   type ScriptsNeeded ConwayEra = AlonzoScriptsNeeded ConwayEra
-
-  consumed = conwayConsumed
 
   getConsumedValue = getConsumedMaryValue