+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Rules/Utxo.hs
@@ -65,7 +65,7 @@ import Cardano.Ledger.Shelley.Rules.Ppup (
   ShelleyPpupPredFailure,
  )
 import Cardano.Ledger.Shelley.Rules.Reports (showTxCerts)
-import Cardano.Ledger.Shelley.UTxO (produced)
+import Cardano.Ledger.Shelley.UTxO (produced, shelleyConsumed)
 import Cardano.Ledger.Slot (SlotNo)
 import Cardano.Ledger.State
 import Cardano.Ledger.TxIn (TxIn)
@@ -514,8 +514,9 @@ validateValueNotConservedUTxO pp utxo certState txBody =
   failureUnless (consumedValue == producedValue) $
     ValueNotConservedUTxO Mismatch {mismatchSupplied = consumedValue, mismatchExpected = producedValue}
   where
-    consumedValue = consumed pp certState utxo txBody
-    producedValue = produced pp certState txBody
+    accounts = certState ^. certDStateL . accountsL
+    consumedValue = shelleyConsumed pp accounts utxo txBody
+    producedValue = produced pp (certState ^. certPStateL) txBody
 
 -- | Ensure there are no `TxOut`s that have less than @minUTxOValue@
 --
@@ -624,7 +625,7 @@ updateUTxOStateNoFees pp utxos txBody certState govState depositChangeEvent txUt
       {- newUTxO = (txins txb ⋪ utxo) ∪ outs txb -}
       newUTxO = utxoWithout `Map.union` unUTxO utxoAdd
       deletedUTxO = UTxO utxoDel
-      totalRefunds = certsTotalRefundsTxBody pp certState txBody
+      totalRefunds = certsTotalRefundsTxBody pp (certState ^. certDStateL . accountsL) txBody
       totalDeposits = certsTotalDepositsTxBody pp certState txBody
       depositChange = totalDeposits <-> totalRefunds
   depositChangeEvent depositChange