+++ libs/cardano-ledger-core/src/Cardano/Ledger/State/UTxO.hs
@@ -244,15 +244,11 @@ class EraTx era => EraUTxO era where
   -- scripts needed for the transaction.
   type ScriptsNeeded era = (r :: Type) | r -> era
 
-  consumed :: PParams era -> CertState era -> UTxO era -> TxBody t era -> Value era
-
   -- | Calculate all the value that is being consumed by the transaction.
   getConsumedValue ::
     PParams era ->
     -- | Function that can lookup current delegation deposits
     (Credential Staking -> Maybe Coin) ->
-    -- | Function that can lookup current drep deposits
-    (Credential DRepRole -> Maybe Coin) ->
     UTxO era ->
     TxBody t era ->
     Value era