+++ eras/dijkstra/impl/CHANGELOG.md
@@ -1,7 +1,8 @@
-# Revision history for cardano-ledger-dijkstra
+# Revision history for `cardano-ledger-dijkstra`
 
 ## 0.3.0.0
 
+* Add `dijkstraConsumed` and `validateValueNotConservedUTxO`
 * Make `requiredTopLevelGuards` available on the top-level transaction body:
   - Change the type of `requiredTopLevelGuardsL` to `Lens' (TxBody l era) (Map (Credential Guard) (StrictMaybe (Data era)))`
   - Change the type of `requiredTopLevelGuardsDijkstraTxBodyRawL` to `Lens' (DijkstraTxBodyRaw l era) (Map (Credential Guard) (StrictMaybe (Data era)))`